move methods to new file without any changes, fix tests import paths
