+++ packages/webdriverio/src/commands/element/getHTML.js
@@ -30,7 +30,7 @@
  */
 
 import { ELEMENT_KEY } from '../../constants'
-import { getBrowserObject } from '../../utils'
+import { getBrowserObject } from '../../find-strategy'
 import getHTMLScript from '../../scripts/getHTML'
 
 export default function getHTML (includeSelectorTag = true) {
