+++ packages/webdriverio/src/commands/element/getSize.js
@@ -27,7 +27,7 @@
  *
  */
 
-import { getElementRect } from '../../utils'
+import { getElementRect } from '../../find-strategy'
 
 export default async function getSize(prop = null) {
     let rect = {}
